+++ lib/pro_video_editor.dart
@@ -4,16 +4,16 @@ export 'core/models/video/progress_model.dart';
 export 'core/models/video/editor_video_model.dart';
 export 'core/models/video/export_transform_model.dart';
 export 'core/models/video/render_video_model.dart';
-export 'core/models/video/video_clip_model.dart';
+export 'core/models/video/video_segment_model.dart';
 export 'core/models/video/video_metadata_model.dart';
 export 'core/models/video/video_quality_preset.dart';
 export 'core/models/video/video_quality_config.dart';
-export 'core/models/video/render_exceptions.dart';
-export 'pro_video_editor_platform_interface.dart';
+export 'core/models/exceptions/render_exceptions.dart';
+export 'core/platform/platform_interface.dart';
 export 'shared/utils/converters.dart';
 
 /// Thumbnails
-export 'core/models/thumbnail/key_frames_configs.model.dart';
-export 'core/models/thumbnail/thumbnail_box_fit.model.dart';
-export 'core/models/thumbnail/thumbnail_configs.model.dart';
-export 'core/models/thumbnail/thumbnail_format.model.dart';
+export 'core/models/thumbnail/key_frames_configs_model.dart';
+export 'core/models/thumbnail/thumbnail_box_fit_model.dart';
+export 'core/models/thumbnail/thumbnail_configs_model.dart';
+export 'core/models/thumbnail/thumbnail_format_model.dart';