+++ test/pro_video_editor_method_channel_test.dart
@@ -2,8 +2,8 @@ import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
+import 'package:pro_video_editor/core/platform/native_method_channel.dart';
 import 'package:pro_video_editor/pro_video_editor.dart';
-import 'package:pro_video_editor/pro_video_editor_method_channel.dart';
 
 import 'pro_video_editor_method_channel_test.mocks.dart';
 
@@ -11,7 +11,7 @@ import 'pro_video_editor_method_channel_test.mocks.dart';
   EditorVideo,
   ThumbnailConfigs,
   KeyFramesConfigs,
-  RenderVideoModel,
+  VideoRenderData,
 ])
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();